requirements.txt
matplotlib==3.4.3
numpy==1.21.4
numpy==1.21.2
Pillow==8.4.0
Pillow==10.0.0
vip_hci==1.3.5
vip_hci==1.3.1
